requirements.txt
tensorflow-directml-plugin
numpy<2
pandas